+++ src/codec_avif_libheif.cc
@@ -33,7 +33,7 @@
 #include "src/timer.h"
 
 #if defined(HAS_WEBP2)
-#include "third_party/libwebp2/src/wp2/base.h"
+#include "src/wp2/base.h"
 #endif
 
 #if defined(HAS_HEIF)